Go back to using Fixed version of Log2 over the Fixed128 version
